api/pyproject.toml
[project]
name = "bluestart-backend2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi[standard]>=0.127.1",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==3.2.2",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.45",
    "asyncpg>=0.31.0",
]
